% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/laQCA.R
\name{laQCA}
\alias{laQCA}
\title{Lieberson Assessment}
\usage{
laQCA(mod, sim = 2000, verbose = T, include = c(""), row.dom = F,
  all.sol = F, omit = c(), dir.exp = c())
}
\arguments{
\item{mod}{name of the model object; the table of solutions for an application of QCA}

\item{sim}{the number of simulations the laQCA function should run. The default is set at 2000}
}
\value{
The proportion of the QCA solutions set that is random
}
\description{
Post-QCA, this function takes in and simulates the parameters of QCA solutions to give a proportion of randomness in solutions.
}
\examples{
data(rallies)
P<-rallies$P
R<-rallies$R
C<-rallies$C
U<-rallies$U

conditions<-data.frame(P,R,C,U)
truth<-truthTable(conditions,outcome="P",sort.by="incl",incl.cut1=0.7,show.cases=TRUE)
truth
mod1 <- eqmcc(truth,details=TRUE,show.cases=TRUE)
mod1

laQCA(mod1)
}

