% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.ltQCA.R
\name{sim.ltQCA}
\alias{sim.ltQCA}
\title{simulation information}
\usage{
sim.ltQCA(qca.data, outcome = "OUT", conditions = c(""), sim = 10,
  ncut = 2, type = "crisp", inclcut = "", neg.out = F, verbose = T)
}
\value{
nothing
#'
}
\description{
This simulation document works for the package
}

